+++ qtypes/_tree_widget.py
@@ -38,6 +38,11 @@ class TreeWidget(QtWidgets.QTreeWidget):
             self.setItemWidget(item, 1, widget)
             self.children.append(item)
 
+    def clear(self):
+        while self.children:
+            child = self.children.pop(0)
+            self.takeTopLevelItem(0)
+
     def insert(self, index, item):
         if index < 0:
             index += self.topLevelItemCount()